requirements.txt
python-3.7.4
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.12
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
